Updating zephyr version:
To prevent some IDE freezes and to speed things up, the
logging subsystem now detects if some preprocessor macros
are defined at build time.

If so, all the LOG_x macro are replaced with no-op function.

The no-op function keeps the printf format attribute to
keep the IDE checking for incorrect formatted strings