Modified nRF's doc generation to include Intellinium's documentation
This also modifies the theme of the documentation to add the Intellinium logo and give flexibility for future modifications.

This should not cause too much conflict with future documentation updates from Nordic. In case a new documentation is added by Nordic, its theme should be updated to remain consistent with the rest. For that, just follow the modifications on:
    * nrf/doc/[the_new_doc]/conf.py
    * nrf/doc/CMakeLists.txt (just add the ITL_LIB_BASE environment variable to the python execution)

All the Intellinium documentation content is located under ${ITL_LIB_BASE}/doc