dfu: FOTA over UART
This PR adds the possibility to forward an update image received from
an LTE connection to another chip by using UART communication.

The image is sent progressively and the download of the image is
cancelled if one of the packet is not received by the other side of the
UART bus.

This PR requires no change on the MCUBoot side, but require the signed
update image to be wrapped to be both compatible with the uart dfu type
and compatible with all the MCUBoot versions.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>